requirements.txt
torchio==0.18.83 
monai==0.8.1
pytorch-lightning==1.6.3
gdown==4.4.0
matplotlib==3.2.2
pandas==1.3.5
seaborn==0.11.2
ipykernel